+++ ChargeBee/Api/ApiException.cs
@@ -1,61 +1,65 @@
 ﻿using System;
 using System.Net;
 using System.Runtime.Serialization;
+using System.Collections.Generic;
 
 namespace ChargeBee.Api
 {
     public class ApiException : ApplicationException
     {
-        public ApiException(SerializationInfo info, StreamingContext context)
+
+		private string ErrorType = "";
+		private string ErrorParam = "";
+
+		public ApiException (HttpStatusCode httpStatusCode, Dictionary<string, string> errorResp)
+			: base (errorResp ["message"])
         {
-            if (info != null)
-            {
-                foreach (var item in info)
-                {
-                    switch (item.Name)
-                    {
-                        case "error_code":
-                            ApiCode = item.Value.ToString();
-                            break;
-                        case "error_param":
-                            Parameter = item.Value.ToString();
-                            break;
-                        case "error_msg":
-                            ApiMessage = item.Value.ToString();
-                            break;
-                        default:
-                            break;
-                    }
-                }
-            }
+			this.HttpStatusCode = httpStatusCode;
+			errorResp.TryGetValue ("type", out ErrorType);
+			this.ApiErrorCode = errorResp ["api_error_code"];
+
+			errorResp.TryGetValue("param", out ErrorParam);
+
+			//Deprecated fields.
+			this.ApiCode = errorResp ["error_code"];
+			this.ApiMessage = errorResp ["error_msg"];
         }
 
-        public override void GetObjectData(SerializationInfo info, StreamingContext context)
-        {
-            base.GetObjectData(info, context);
+        public HttpStatusCode HttpStatusCode { get; set; }
 
-            if (info != null)
-            {
-                info.AddValue("error_code", ApiCode);
-                info.AddValue("error_param", Parameter);
-                info.AddValue("error_msg", ApiMessage);
-            }
-        }
+		public string Type { 
+			get {
+				return this.ErrorType;
+			}
+		}
 
-        public HttpStatusCode HttpCode { get; set; }
+		public string ApiErrorCode { get; set; }
 
-        public string ApiCode { get; set; }
+        public string Param { 
+			get {
+				return this.ErrorParam;
+			}
+		}
 
-        public string Parameter { get; set; }
+        [System.Obsolete("Use HttpStatusCode")]
+        public HttpStatusCode HttpCode { 
+			get {
+				return HttpStatusCode;
+			}
+		}
 
+        [System.Obsolete("Use Code")]
+		public string ApiCode { get; set; }
+
+        [System.Obsolete("Use Param")]
+		public string Parameter { 
+			get { 
+				return Param;
+			} 
+		}
+
+        [System.Obsolete("Use Message")]
         public string ApiMessage { get; set; }
 
-        public override string Message
-        {
-            get
-            {
-                return ApiMessage;
-            }
-        }
     }
 }