+++ ChargeBee/Internal/ResultBase.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 
+using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
 using ChargeBee.Models;
@@ -15,7 +16,14 @@ namespace ChargeBee.Internal
         internal ResultBase(string json)
         {
             if (!String.IsNullOrEmpty(json))
-                m_jobj = JToken.Parse(json);
+			{
+				try 
+				{
+                	m_jobj = JToken.Parse(json);
+				} catch(JsonException e){
+					throw new SystemException("Not in JSON format. Probably not a ChargeBee response. \n " + json, e);
+				}
+			}
         }
 
         internal ResultBase(JToken jobj)