+++ ChargeBee/Models/Enums/EventTypeEnum.cs
@@ -48,6 +48,7 @@ namespace ChargeBee.Models.Enums
         [Description("invoice_created")]
         InvoiceCreated,
 
+
         [Description("payment_succeeded")]
         PaymentSucceeded,